Add small optimizations to 0058

+++ 0058 - Length of Last Word/solution.c
@@ -3,17 +3,17 @@
 
 int lengthOfLastWord(char * s){
 
-    int indexOfFinalFirstChar = -1;
+    short indexOfFinalFirstChar = -1;
     unsigned char lastCharWasSpace = 1;
 
-    for (int i = 0; s[i] != 0; ++i) {
+    for (short i = 0; s[i] != 0; ++i) {
         if (lastCharWasSpace && s[i] != ' ')
             indexOfFinalFirstChar = i;
 
         lastCharWasSpace = s[i] == ' ';
     }
 
-    int i;
+    short i;
     for (i = indexOfFinalFirstChar; s[i] != 0; ++i) {
         if (s[i] == ' ')
             break;
